added 'core' module
