+++ app/src/main/java/com/example/android/architecture/blueprints/todoapp/taskdetail/TaskDetailContract.java
@@ -16,8 +16,8 @@
 
 package com.example.android.architecture.blueprints.todoapp.taskdetail;
 
-import com.example.android.architecture.blueprints.todoapp.BasePresenter;
-import com.example.android.architecture.blueprints.todoapp.BaseView;
+import com.example.core.BasePresenter;
+import com.example.core.BaseView;
 
 /**
  * This specifies the contract between the view and the presenter.
