+++ app/src/main/java/com/example/android/architecture/blueprints/todoapp/tasks/TasksContract.java
@@ -18,8 +18,8 @@ package com.example.android.architecture.blueprints.todoapp.tasks;
 
 import android.support.annotation.NonNull;
 
-import com.example.android.architecture.blueprints.todoapp.BasePresenter;
-import com.example.android.architecture.blueprints.todoapp.BaseView;
+import com.example.core.BasePresenter;
+import com.example.core.BaseView;
 import com.example.data.Task;
 
 import java.util.List;
