+++ core/src/main/java/com/example/core/BasePresenter.java
@@ -14,10 +14,12 @@
  * limitations under the License.
  */
 
-package com.example.android.architecture.blueprints.todoapp;
+package com.example.core;
 
-public interface BaseView<T> {
+public interface BasePresenter {
 
-    void setPresenter(T presenter);
+    void subscribe();
+
+    void unsubscribe();
 
 }
