+++ settings.gradle
@@ -1,1 +1,1 @@
-include ':app', ':data', ':utils'
+include ':app', ':data', ':utils', ':core'
